Rewrite navigation panel
